--- a/asconfig.json
+++ b/asconfig.json
@@ -10,11 +10,11 @@
     "release": {
       "binaryFile": "build/optimized.wasm",
       "textFile": "build/optimized.wat",
-      "sourceMap": true,
       "optimizeLevel": 3,
-      "shrinkLevel": 1,
-      "converge": false,
-      "noAssert": false,
+      "shrinkLevel": 0,
+      "converge": true,
+      "noAssert": true,
+      "sourceMap": true,
       "exportRuntime": true
     }
   },