--- a/assembly/index.ts
+++ b/assembly/index.ts
@@ -1,36 +1,38 @@
 // The entry file of your WebAssembly module.
 
+export const Int32Array_ID = idof<Int32Array>();
+export const Float64Array_ID = idof<Float64Array>();
+
 export function add(a: i32, b: i32): i32 {
   return a + b;
 }
 
-export function factorial(i: f64): f64 {
-  if (i == 0) return 1;
-  return i * factorial(i - 1);
+export function factorial(i: i32): i32 {
+  return i == 0 ? 1 : i * factorial(i - 1);
 }
 
 export function squareArray(arr: Int32Array): Int32Array {
-  const result = new Int32Array(arr.length);
-  for (let i = 0; i < arr.length; ++i) {
-    unchecked((result[i] = arr[i] * arr[i]));
+  const len = arr.length;
+  const result = new Int32Array(len);
+  for (let i = 0; i < len; ++i) {
+    const e = unchecked(arr[i]);
+    unchecked(result[i] = e * e);
   }
   return result;
 }
 
-export const Int32Array_ID = idof<Int32Array>();
 
 export function calcSinLookup(): Float64Array {
   const max = 6283;
   const result = new Float64Array(max);
 
   for (let i = 0; i < max; ++i) {
-    result[i] = Math.sin(i * 0.001);
+    unchecked(result[i] = Math.sin(i * 0.001));
   }
 
   return result;
 }
 
-export const Float64Array_ID = idof<Float64Array>();
 
 declare namespace test {
   @external("test", "importCallback")
@@ -38,7 +40,7 @@ declare namespace test {
 }
 
 export function testImport(n: i32): i32 {
-  let result: i32 = 0;
+  let result = 0;
   for (let i = 1; i <= n; ++i) {
     result = test.importCallback(result, i);
   }