I would like suggest some improvements. Also I guess better use same data structures on JS and AS side, so I changed usual arrays to typed arrays on JS side.

Note basically for such simple benchmarks on JS side you're usually measure interop overhead instead actual wasm performance but you probably are aware of this.

Anther suggestion. I recommend exclude array creation for `calcSinLookup` or create it globally and once and rewrite it every benchmark stage. In this case you could compare JS and AS math more properly

master (before):
```
Benchmarking add:
AssemblyScript x 152,771,389 ops/sec ±0.67% (89 runs sampled)
JavaScript x 872,744,847 ops/sec ±1.10% (85 runs sampled)
JavaScript is faster

Benchmarking factorial:
AssemblyScript x 11,824,147 ops/sec ±1.04% (92 runs sampled)
JavaScript x 6,813,119 ops/sec ±0.64% (94 runs sampled)
AssemblyScript is faster

Benchmarking squareArray:
AssemblyScript x 305,892 ops/sec ±1.18% (91 runs sampled)
JavaScript x 2,167,033 ops/sec ±0.79% (91 runs sampled)
JavaScript is faster

Benchmarking calcSinLookup:
AssemblyScript x 12,320 ops/sec ±0.65% (91 runs sampled)
JavaScript x 12,372 ops/sec ±0.67% (95 runs sampled)
JavaScript,AssemblyScript is faster

Benchmarking importCallback:
AssemblyScript x 29,308 ops/sec ±0.35% (94 runs sampled)
JavaScript x 419,580 ops/sec ±0.41% (96 runs sampled)
JavaScript is faster
```

with this PR (after):
```
Benchmarking add:
AssemblyScript x 156,504,221 ops/sec ±0.64% (91 runs sampled)
JavaScript x 893,287,993 ops/sec ±1.49% (87 runs sampled)
JavaScript is faster

Benchmarking factorial:
AssemblyScript x 12,210,579 ops/sec ±0.84% (95 runs sampled)
JavaScript x 6,838,418 ops/sec ±0.71% (94 runs sampled)
AssemblyScript is faster

Benchmarking squareArray:
AssemblyScript x 458,934 ops/sec ±1.25% (88 runs sampled)
JavaScript x 299,228 ops/sec ±3.15% (85 runs sampled)
AssemblyScript is faster

Benchmarking calcSinLookup:
AssemblyScript x 11,483 ops/sec ±1.19% (86 runs sampled)
JavaScript x 10,091 ops/sec ±0.58% (92 runs sampled)
AssemblyScript is faster

Benchmarking importCallback:
AssemblyScript x 29,004 ops/sec ±0.81% (94 runs sampled)
JavaScript x 419,500 ops/sec ±0.32% (98 runs sampled)
JavaScript is faster
```