@LeanyLabs 2021-02-07
Thanks for optimizing the tests!

Yeah, we were more testing interop to see at which point WASM makes sense over plain JavaScript. 

The calculations are faster and consistent, but if they aren't that complicated and it involves passing data back and forth it might destroy the benefits.

I've updated the squareArray JS test with pretty much the exact AS version, and it's still faster in JS:

```
Benchmarking squareArray:
AssemblyScript x 432,270 ops/sec ±1.32% (87 runs sampled)
JavaScript x 822,068 ops/sec ±2.31% (85 runs sampled)
JavaScript is faster
```

Love the project btw. Hoping we'll build something with AS in production soon ;).